move LiveBlock to a standalone component
